Move `VerifyPluginWorkAction` to `workers` package and refactor related usage in `VerifyPluginTask`. Minor adjustments
